+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
+- Validate checkpoint block before adding it to the acceptance queue
 
 ## [v2.25.2] 2021-07-09
 ### Fixed